changes to the include files

+++ LinearController/drivers/adc.h
@@ -5,15 +5,16 @@
  *  Author: hasna
  */ 
 
+#include <avr/io.h>
+#include <stdio.h>
+
 
 #ifndef ADC_H_
 #define ADC_H_
 
-#include <stdint.h>
-
 void adc_init();
 uint16_t adc_read();
-float adc_calculate();
+uint8_t adc_calculate();
